Renamed components for more clarity and consistency
