+++ frontend/app/[locale]/alerts/page.tsx
@@ -1,6 +1,6 @@
 /* * */
 
-import AlertsPage from '@/components/alerts/Page';
+import AlertsList from '@/components/alerts/AlertsList';
 import { AlertsListContextProvider } from '@/contexts/AlertsList.context';
 
 /* * */
@@ -8,7 +8,7 @@ import { AlertsListContextProvider } from '@/contexts/AlertsList.context';
 export default function Page() {
 	return (
 		<AlertsListContextProvider>
-			<AlertsPage />
+			<AlertsList />
 		</AlertsListContextProvider>
 	);
 }
