+++ frontend/components/alerts/AlertsDetail/styles.module.css
